requirements.txt
pydub
openai-whisper
pyannote.audio